pyproject.toml
[project]
name = "hlamovshchik-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "ech0p1ng",email = "alexey2003petrov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "mypy (>=1.19.0,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "httpx (>=0.28.1,<0.29.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "watchfiles (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "minio (>=7.2.20,<8.0.0)",
    "logger (>=1.4,<2.0)",
    "debugpy (>=1.8.19,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
